ai_tutor/backend/requirements.txt
fastapi
uvicorn
PyJWT
python-multipart
bcrypt
passlib[bcrypt]
oracledb
pydantic
numpy
